pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lumena-scanner"
version = "1.0.0"
description = "Custom AI code scanner powered by Lumena's Flame Protocol — Driftprint, Vaultwatch, Token Shield"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "KingofLumena", email = "contact@lumena.io"}
]
keywords = ["security", "scanner", "ai", "secrets", "vault"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "colorama>=0.4.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
lumena = "lumena.cli:main"

[project.urls]
Homepage = "https://github.com/KingofLumena/lumena-scanner"
Repository = "https://github.com/KingofLumena/lumena-scanner"
Issues = "https://github.com/KingofLumena/lumena-scanner/issues"

[tool.setuptools]
packages = ["lumena", "lumena.detectors"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
